requirements.txt
flask
torch
werkzeug
opencv-python
pillow
ultralytics
flask-cors
gunicorn 